ENEL 645/A2/requirements.txt
torch
numpy
glob2
matplotlib
Pillow
scikit-learn
torchvision
torchmetrics
pytorch-lightning
wandb